@@ -1,8 +1,6 @@
-module.exports = {
-  theme: {
-    color: {
-      primary: '#45b29a',
-      secondary: 'rgb(100, 255, 218)'
-    }
+export const theme = {
+  color: {
+    primary: '#45b29a',
+    secondary: 'rgb(100, 255, 218)'
   }
 }
