@@ -1,58 +1,58 @@
-module.exports = {
-  lamboplace: {
-    name: 'Lamboplace',
-    personal: false,
-    madeAt: 'Witty Data',
-    tools: [
-      '#Nextjs',
-      '#Reactjs',
-      '#React-Native',
-      '#Magento'
-    ],
-    img: '/img/lamboplace.webp',
-    github: false,
-    link: 'https://lamboplace.com'
-  },
-  ikang: {
-    name: 'Ikang',
-    personal: false,
-    madeAt: 'Witty Data',
-    tools: [
-      '#Nextjs',
-      '#React-Native',
-      '#Antd',
-      '#MongoDB'
-    ],
-    img: '/img/ikang.webp',
-    github: false,
-    link: 'https://imaninsani.com'
-  },
-  saraponic: {
-    name: 'SaraPonics',
-    personal: true,
-    madeAt: 'Final Year Project',
-    tools: [
-      '#C++',
-      '#Arduino',
-      '#RaspberryPi',
-      '#Node-Red'
-    ],
-    img: '/img/saraponic.webp',
-    github: true,
-    link: 'https://github.com/wadzee/automated-hydroponics'
-  },
-  webportfolio: {
-    name: 'Web Portfolio',
-    personal: true,
-    madeAt: 'Personal Project',
-    tools: [
-      '#Nextjs',
-      '#Antd',
-      '#Google Analytics',
-      '#Lambda'
-    ],
-    img: '/img/portfolio.webp',
-    github: true,
-    link: 'https://github.com/wadzee/web-portfolio'
-  }
+const lamboplace = {
+  name: 'Lamboplace',
+  personal: false,
+  madeAt: 'Witty Data',
+  tools: [
+    '#Nextjs',
+    '#Reactjs',
+    '#React-Native',
+    '#Magento'
+  ],
+  img: '/img/lamboplace.webp',
+  github: false,
+  link: 'https://lamboplace.com'
 }
+const ikang = {
+  name: 'Ikang',
+  personal: false,
+  madeAt: 'Witty Data',
+  tools: [
+    '#Nextjs',
+    '#React-Native',
+    '#Antd',
+    '#MongoDB'
+  ],
+  img: '/img/ikang.webp',
+  github: false,
+  link: 'https://imaninsani.com'
+}
+const saraponic = {
+  name: 'SaraPonics',
+  personal: true,
+  madeAt: 'Final Year Project',
+  tools: [
+    '#C++',
+    '#Arduino',
+    '#RaspberryPi',
+    '#Node-Red'
+  ],
+  img: '/img/saraponic.webp',
+  github: true,
+  link: 'https://github.com/wadzee/automated-hydroponics'
+}
+const webportfolio = {
+  name: 'Web Portfolio',
+  personal: true,
+  madeAt: 'Personal Project',
+  tools: [
+    '#Nextjs',
+    '#Antd',
+    '#Google Analytics',
+    '#Lambda'
+  ],
+  img: '/img/portfolio.webp',
+  github: true,
+  link: 'https://github.com/wadzee/web-portfolio'
+}
+
+export default { webportfolio, saraponic, ikang, lamboplace}
