@@ -1,23 +1,23 @@
-module.exports = {
-  skills: {
-    front: [
-      'JavaScript (ES6+)',
-      'React Native',
-      'Reactjs',
-      'Nextjs'
-    ],
-    back: [
-      'Node.js',
-      'PostgreSQL',
-      'MongoDB',
-      'Firebase'
-    ]
-  },
-  work: {
-    witty: [
-        'Worked with a team of senior developer to build e-commerce website platform such as LamboPlace and Ikang',
-        'Develop for both website and mobile application using a variety of different language, frameworks, and content management system such as JavaScript, React, React Native, Nextjs, Magento',
-        'Develop and create own API for each application backend with MongoDB & PostgreSQL'
-      ]
-  }
+const skills = {
+  front: [
+    'JavaScript (ES6+)',
+    'React Native',
+    'Reactjs',
+    'Nextjs'
+  ],
+  back: [
+    'Node.js',
+    'PostgreSQL',
+    'MongoDB',
+    'Firebase'
+  ]
 }
+const work = {
+  witty: [
+    'Worked with a team of senior developer to build e-commerce website platform such as LamboPlace and Ikang',
+    'Develop for both website and mobile application using a variety of different language, frameworks, and content management system such as JavaScript, React, React Native, Nextjs, Magento',
+    'Develop and create own API for each application backend with MongoDB & PostgreSQL'
+  ]
+}
+
+export default { skills, work }
